+++ src/dotnet/commands/dotnet-add/dotnet-add-p2p/LocalizableStrings.cs
@@ -1,0 +1,11 @@
+namespace Microsoft.DotNet.Tools.Add.ProjectToProjectReference
+{
+    internal class LocalizableStrings
+    {
+        public const string ReferenceIsInvalid = "Reference `{0}` is invalid.";
+
+        public const string SpecifyAtLeastOneReferenceToAdd = "You must specify at least one reference to add. Please run dotnet add --help for more information.";
+
+        public const string ProjectAlreadyHasAReference = "Project {0} already has a reference `{1}`.";
+    }
+}
