+++ src/dotnet/commands/dotnet-remove/dotnet-remove-p2p/Program.cs
@@ -5,9 +5,9 @@ using Microsoft.DotNet.Cli.CommandLine;
 using Microsoft.DotNet.Cli.Utils;
 using System.Collections.Generic;
 
-namespace Microsoft.DotNet.Tools.Add.ProjectToProjectReference
+namespace Microsoft.DotNet.Tools.Remove.ProjectToProjectReference
 {
-    public class AddProjectToProjectReferenceCommand
+    public class RemoveProjectToProjectReferenceCommand
     {
         public static int Run(string[] args)
         {
@@ -15,11 +15,11 @@ namespace Microsoft.DotNet.Tools.Add.ProjectToProjectReference
 
             CommandLineApplication app = new CommandLineApplication(throwOnUnexpectedArg: false)
             {
-                Name = "dotnet add p2p",
-                FullName = ".NET Add Project to Project (p2p) reference Command",
-                Description = "Command to add project to project (p2p) reference",
+                Name = "dotnet remove p2p",
+                FullName = ".NET Remove Project to Project (p2p) reference Command",
+                Description = "Command to remove project to project (p2p) reference",
                 AllowArgumentSeparator = true,
-                ArgumentSeparatorHelpText = "Project to project references to add"
+                ArgumentSeparatorHelpText = "Project to project references to remove"
             };
 
             app.HelpOption("-h|--help");
@@ -32,39 +32,29 @@ namespace Microsoft.DotNet.Tools.Add.ProjectToProjectReference
 
             CommandOption frameworkOption = app.Option(
                 "-f|--framework <FRAMEWORK>",
-                "Add reference only when targetting a specific framework",
+                "Remove reference only when targetting a specific framework",
                 CommandOptionType.SingleValue);
-
-            CommandOption forceOption = app.Option(
-                "--force", 
-                "Add reference even if it does not exist, do not convert paths to relative",
-                CommandOptionType.NoValue);
 
             app.OnExecute(() => {
                 if (string.IsNullOrEmpty(projectArgument.Value))
                 {
-                    throw new GracefulException(LocalizableStrings.RequiredArgumentNotPassed, "<Project>");
+                    throw new GracefulException(CommonLocalizableStrings.RequiredArgumentNotPassed, "<Project>");
                 }
 
                 var msbuildProj = MsbuildProject.FromFileOrDirectory(projectArgument.Value);
 
                 if (app.RemainingArguments.Count == 0)
                 {
-                    throw new GracefulException(LocalizableStrings.SpecifyAtLeastOneReferenceToAdd);
+                    throw new GracefulException(LocalizableStrings.SpecifyAtLeastOneReferenceToRemove);
                 }
 
                 List<string> references = app.RemainingArguments;
-                if (!forceOption.HasValue())
-                {
-                    MsbuildProject.EnsureAllReferencesExist(references);
-                    msbuildProj.ConvertPathsToRelative(ref references);
-                }
                 
-                int numberOfAddedReferences = msbuildProj.AddProjectToProjectReferences(
+                int numberOfRemovedReferences = msbuildProj.RemoveProjectToProjectReferences(
                     frameworkOption.Value(),
                     references);
 
-                if (numberOfAddedReferences != 0)
+                if (numberOfRemovedReferences != 0)
                 {
                     msbuildProj.Project.Save();
                 }
